+++ .github/workflows/module_tests.yml
@@ -56,5 +56,5 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with unittest
       run: |
-        python -m unittest tests/module.py
+        python -m unittest tests/test_module.py
