--- dbtest/__init__.py
@@ -1,3 +1,0 @@
-from dbtest.unittest import *
-from dbtest.postgresql import *
-from dbtest.utils import *
